check stack's max before malloc

+++ include/gtstack.h
@@ -11,7 +11,7 @@ extern "C" {
 #include "gttypes.h"
 typedef struct gt_stack gt_stack;
 
-GT_API gt_stack* gt_stack_create(int);
+GT_API gt_stack* gt_stack_create(size_t);
 GT_API int gt_stack_push(gt_stack*, void*);
 GT_API int gt_stack_pop(gt_stack*, void**);
 GT_API void gt_stack_destroy(gt_stack**);
